Airtable

Airtable.Credential

Airtable.get

Airtable.patch

Airtable.post

Airtable.put

Airtable.query

Airtable.request

Airtable.Credential — Type

 Credential(; api_key)

 A credentials and identity object for Trello.

If the apikey or apitoken are not provided, they will be read from the AIRTABLE_KEY environment variable. Go to Airtable account settings to aquire your credentials.

Airtable.get — Method

Airtable.get(cred::Credential, path::AbstractString; query_kwargs...)

Shorthand for Airtable.request("GET", cred, path; query_kwargs)

Airtable.patch — Method

Airtable.patch(cred::Credential, path::AbstractString; query_kwargs...)

Shorthand for Airtable.request("PATCH", cred, path; query_kwargs)

Airtable.post — Method

Airtable.post(cred::Credential, path::AbstractString; query_kwargs...)

Shorthand for Airtable.request("POST", cred, path; query_kwargs)

Airtable.put — Method

Airtable.put(cred::Credential, path::AbstractString; query_kwargs...)

Shorthand for Airtable.request("PUT", cred, path; query_kwargs)

Airtable.query — Method

Airtable.query(cred::Credential, path::AbstractString; query_kwargs...)

Shorthand for a "GET" request that handles continuation and rate-limiting.

The Airtable API will return a maximum of 100 records per requests, and only allows 5 requests / sec. This function uses the offset field returned as part of a requst that does not contain all possible records to make additional requests after pausing 0.21 seconds in between.

Required arguments:

cred: an Airtable.Credential containing your API key

path: the endpoint of your Airtable base. See https://airtable.com/api for details

Query parameters are in the form of keyword arguments, eg filterByFormla = "NOT({Name} = '')", maxRecords=2. See Airtable API reference for more information.

Airtable.request — Method

Airtable.request(method::AbstractString, cred::Credential, path::AbstractString; query_kwargs...)

Make a request to the Airtable API.

Required arguments:

method: one of "GET", "PUT", "POST", or "PATCH",

cred: an Airtable.Credential containing your API key

path: the endpoint of your Airtable base. See https://airtable.com/api for details

Query parameters are in the form of keyword arguments, eg filterByFormla = "NOT({Name} = '')", maxRecords=2. See Airtable API reference for more information.