Interface

A number of functions have been added to make it a bit easier to interact with the Airtable API. These are built on top of the low-level interface, which is just a thin wrapper around the HTTP.jl functions.

Airtable.query function

If you are interested in retrieving records, your best bet is the Airtable.query function. Because the airtable API will only return 100 records per request[1], and only allows 5 requests/sec, query() automatically handles the "paging", pausing for 0.2 seconds in between each request so that you won't hit your limit.

At minimum, query() requires an AirTable (note the capital 'T'), which requires that you know the base name and table name for your air table. If you only pass the table, you will retrieve all of the records for that table.

julia> tab = AirTable("Table 1", AirBase("appphImnhJO8AXmmo"))
AirTable("Table 1")

julia> Airtable.query(tab)
3-element Vector{AirRecord}:
 AirRecord("recMc1HOSIxQPJyyc", AirTable("Table 1"), (Name = "Record 1", Notes = "Keep this\n", Status = "Todo", Keep = true))
 AirRecord("recMwT4P4tKlSLJoH", AirTable("Table 1"), (Name = "Record 2", Notes = "Also keep this", Status = "In progress", Keep = t
rue))
 AirRecord("recSStgr3yJnQc2Wg", AirTable("Table 1"), (Name = "Record 3", Status = "Done", Keep = true))

Here, the Credential is used automatically from the AIRTABLE_KEY environmental variable. As you can see, unlike the low-level functions that return JSON3.Objects, query() returns AirRecords

You may also pass additional query parameters (passed as keyword arguments) to filter, or otherwise modify the query.

julia> Airtable.query(tab; filterByFormula="{Status} = 'Todo'")
1-element Vector{AirRecord}:
 AirRecord("recMc1HOSIxQPJyyc", AirTable("Table 1"), (Name = "Record 1", Notes = "Keep this\n", Status = "Todo", Keep = true))

julia> Airtable.query(tab; filterByFormula="{Status} != 'Todo'")
2-element Vector{AirRecord}:
 AirRecord("recMwT4P4tKlSLJoH", AirTable("Table 1"), (Name = "Record 2", Notes = "Also keep this", Status = "In progress", Keep = t
rue))
 AirRecord("recSStgr3yJnQc2Wg", AirTable("Table 1"), (Name = "Record 3", Status = "Done", Keep = true))

For more information about usable query parameters, refer to the airtable documentation.

Airtable.AirBase — Type

AirBase(id::String)

A wrapper for an Airtable Base, containing its unique identifier. The ID can be identified from the URL of the base (the part right after airtable.com), or by clicking HELP -> API documentation from within your base.

Airtable.AirTable — Type

AirTable(id::String, ::AirBase)

A wrapper for an Airtable Table, containing its unique identifier or name, and parent AirBase.

Airtable.AirRecord — Type

AirRecord(id::String, table::AirTable, fields::NamedTuple)

A wrapper for an Airtable Record, containing its unique identifier, parent AirTable, and values for any stored fields in a NamedTuple.

Typically, you won't crete these on your own, but they will be returned from API queries.

Field values can be accessed using getindex.

Missing docstring.

Missing docstring for query. Check Documenter's build log for details.

Add/Update Records

I haven't actually figured this out yet 🤔. If you want to help, let me know!

1This is the default, you can change this with the pageSize parameter, but 100 is the maximum.